+++ src/ui/Game/XP.js
@@ -1,0 +1,15 @@
+const React = require('react')
+
+class XP extends React.Component { 
+    constructor() {
+        super()
+    }
+
+    render() {
+        console.log(this.props.replay)
+        return null
+    }
+
+}
+
+module.exports = XP